requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
idna==3.4
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
joblib==1.3.1
jupyter_client==8.2.0
jupyter_core==5.3.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.25.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.5.3
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
redis==4.5.5
requests==2.31.0
rich==13.4.2
scikit-learn==1.3.0
scipy==1.11.1
shellingham==1.5.0.post1
six==1.16.0
spotipy==2.23.0
stack-data==0.6.2
threadpoolctl==3.2.0
tornado==6.3.2
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
